Crate task_maker_diagnostics

This crate contains the code for producing human-friendly and machine-friendly diagnostic messages.

Structs

CodeSpan

A CodeSpan represent a slice of code.

Diagnostic

A Diagnostic is a message, with some extra information attached, such as the message level, additional information about what happened or some help on how to fix the issue.

DiagnosticContext

The context that contains all the emitted diagnostic messages.

Enums

DiagnosticLevel

The level of the message.